Move controller package to infrastructure package
