+++ admin-site/admin-site-back/src/main/java/eu/adainius/newsfocused/admin/site/back/config/JerseyConfig.java
@@ -6,9 +6,9 @@ import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.servlet.ServletProperties;
 import org.springframework.context.annotation.Configuration;
 
-import eu.adainius.newsfocused.admin.site.back.controller.ApiController;
-import eu.adainius.newsfocused.admin.site.back.controller.IndexController;
-import eu.adainius.newsfocused.admin.site.back.controller.auth.AuthenticationController;
+import eu.adainius.newsfocused.admin.site.back.infrastructure.controller.ApiController;
+import eu.adainius.newsfocused.admin.site.back.infrastructure.controller.IndexController;
+import eu.adainius.newsfocused.admin.site.back.infrastructure.controller.auth.AuthenticationController;
 
 @Configuration
 public class JerseyConfig extends ResourceConfig {
