+++ admin-site/admin-site-back/src/main/java/eu/adainius/newsfocused/admin/site/back/infrastructure/controller/package-info.java
@@ -5,4 +5,4 @@
  * A generic exception mapper maps all runtime exceptions to a 500 http status code in the response.
  * Client error codes are handled not by exceptions, but by 'kind of' 'errors' returned with Vavr's Either.left
  */
-package eu.adainius.newsfocused.admin.site.back.controller;
+package eu.adainius.newsfocused.admin.site.back.infrastructure.controller;
