+++ admin-site/admin-site-back/src/test/java/eu/adainius/newsfocused/admin/site/back/infrastructure/controller/ApiControllerTest.java
@@ -1,4 +1,4 @@
-package eu.adainius.newsfocused.admin.site.back.controller;
+package eu.adainius.newsfocused.admin.site.back.infrastructure.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchException;
