+++ admin-site/admin-site-back/src/test/java/eu/adainius/newsfocused/admin/site/back/repositories/UserFileBasedJsonRepositoryTest.java
@@ -21,8 +21,8 @@ import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 
-import eu.adainius.newsfocused.admin.site.back.controller.UserDeserializer;
-import eu.adainius.newsfocused.admin.site.back.controller.UserSerializer;
+import eu.adainius.newsfocused.admin.site.back.infrastructure.controller.UserDeserializer;
+import eu.adainius.newsfocused.admin.site.back.infrastructure.controller.UserSerializer;
 import eu.adainius.newsfocused.admin.site.back.domain.User;
 import eu.adainius.newsfocused.admin.site.back.domain.User.Preferences;
